refactor: group-spending 함수 유틸로 분리
